When TCP client sends logs to the TCP+TLS rsyslog server
Because server expect TLS handshake after the TCP handshake but client
side there is a not clear info about what when wrong, and client
make that connection as disconnected and skip that destination
server while sending the logs.

This patch is just a prooof of concept that we can add a counter
per server/destination which will be logged throught the impstats
and it will help us to write more accurate health checks which
detect that log forwording is happening to a spesific target/server.

This counter is per target/server, and will be printed on the client
side.